extract shared functionality to separate packages
