+++ dev/flight-integration/cases/flight_sql.go
@@ -18,20 +18,17 @@ package cases
 
 import (
 	"context"
-	"encoding/binary"
 	"errors"
 	"fmt"
-	"io"
 
 	"github.com/apache/arrow/dev/flight-integration/protocol/flight"
 	"github.com/apache/arrow/dev/flight-integration/protocol/message/org/apache/arrow/flatbuf"
 	"github.com/apache/arrow/dev/flight-integration/scenario"
+	"github.com/apache/arrow/dev/flight-integration/serialize"
 	"github.com/apache/arrow/dev/flight-integration/tester"
-	flatbuffers "github.com/google/flatbuffers/go"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/types/known/anypb"
 )
 
 func init() {
@@ -62,43 +59,19 @@ func init() {
 
 		createPreparedStatementActionType = "CreatePreparedStatement"
 		closePreparedStatementActionType  = "ClosePreparedStatement"
-
-		queryFields = []field{
-			{
-				Name:         "id",
-				Type:         flatbuf.TypeInt,
-				GetTypeTable: intTypeTable(64, true),
-				Nullable:     true,
-				Metadata: map[string]string{
-					"ARROW:FLIGHT:SQL:TABLE_NAME":        "test",
-					"ARROW:FLIGHT:SQL:IS_AUTO_INCREMENT": "1",
-					"ARROW:FLIGHT:SQL:IS_CASE_SENSITIVE": "0",
-					"ARROW:FLIGHT:SQL:TYPE_NAME":         "type_test",
-					"ARROW:FLIGHT:SQL:SCHEMA_NAME":       "schema_test",
-					"ARROW:FLIGHT:SQL:IS_SEARCHABLE":     "1",
-					"ARROW:FLIGHT:SQL:CATALOG_NAME":      "catalog_test",
-					"ARROW:FLIGHT:SQL:PRECISION":         "100",
-				},
-			},
-		}
 	)
 
 	testcases := []struct {
 		Command proto.Message
-		Fields  []field
+		Fields  []serialize.Field
 	}{
 		{
 			Command: &flight.CommandGetCatalogs{},
-			Fields: []field{
-				{Name: "catalog_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-			},
+			Fields:  serialize.SchemaRef.Catalogs,
 		},
 		{
 			Command: &flight.CommandGetDbSchemas{Catalog: &catalog, DbSchemaFilterPattern: &dbSchemaFilterPattern},
-			Fields: []field{
-				{Name: "catalog_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "db_schema_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-			},
+			Fields:  serialize.SchemaRef.DBSchemas,
 		},
 		// {
 		// 	Command: &flight.CommandGetTables{
@@ -123,66 +96,23 @@ func init() {
 				IncludeSchema:          true,
 				TableTypes:             tableTypes,
 			},
-			Fields: []field{
-				{Name: "catalog_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "db_schema_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "table_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "table_type", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "table_schema", Type: flatbuf.TypeBinary, GetTypeTable: binaryTypeTable(), Nullable: false},
-			},
+			Fields: serialize.SchemaRef.TablesWithIncludedSchema,
 		},
 		{
 			Command: &flight.CommandGetTableTypes{},
-			Fields: []field{
-				{Name: "table_type", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-			},
+			Fields:  serialize.SchemaRef.TableTypes,
 		},
 		{
 			Command: &flight.CommandGetPrimaryKeys{Catalog: &catalog, DbSchema: &dbSchema, Table: table},
-			Fields: []field{
-				{Name: "catalog_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "db_schema_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "table_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "column_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "key_sequence", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: false},
-				{Name: "key_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-			},
+			Fields:  serialize.SchemaRef.PrimaryKeys,
 		},
 		{
 			Command: &flight.CommandGetExportedKeys{Catalog: &catalog, DbSchema: &dbSchema, Table: table},
-			Fields: []field{
-				{Name: "pk_catalog_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "pk_db_schema_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "pk_table_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "pk_column_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "fk_catalog_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "fk_db_schema_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "fk_table_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "fk_column_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "key_sequence", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: false},
-				{Name: "fk_key_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "pk_key_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "update_rule", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(8, false), Nullable: false},
-				{Name: "delete_rule", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(8, false), Nullable: false},
-			},
+			Fields:  serialize.SchemaRef.ExportedKeys,
 		},
 		{
 			Command: &flight.CommandGetImportedKeys{Catalog: &catalog, DbSchema: &dbSchema, Table: table},
-			Fields: []field{
-				{Name: "pk_catalog_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "pk_db_schema_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "pk_table_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "pk_column_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "fk_catalog_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "fk_db_schema_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "fk_table_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "fk_column_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "key_sequence", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: false},
-				{Name: "fk_key_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "pk_key_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "update_rule", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(8, false), Nullable: false},
-				{Name: "delete_rule", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(8, false), Nullable: false},
-			},
+			Fields:  serialize.SchemaRef.ImportedKeys,
 		},
 		{
 			Command: &flight.CommandGetCrossReference{
@@ -193,85 +123,15 @@ func init() {
 				FkDbSchema: &fkDbSchema,
 				FkTable:    fkTable,
 			},
-			Fields: []field{
-				{Name: "pk_catalog_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "pk_db_schema_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "pk_table_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "pk_column_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "fk_catalog_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "fk_db_schema_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "fk_table_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "fk_column_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "key_sequence", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: false},
-				{Name: "fk_key_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "pk_key_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "update_rule", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(8, false), Nullable: false},
-				{Name: "delete_rule", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(8, false), Nullable: false},
-			},
+			Fields: serialize.SchemaRef.CrossReference,
 		},
 		{
 			Command: &flight.CommandGetXdbcTypeInfo{},
-			Fields: []field{
-				{Name: "type_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: false},
-				{Name: "data_type", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: false},
-				{Name: "column_size", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: true},
-				{Name: "literal_prefix", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "literal_suffix", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{
-					Name:         "create_params",
-					Type:         flatbuf.TypeList,
-					GetTypeTable: listTypeTable(field{Name: "item", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable()}),
-					Nullable:     true,
-				},
-				{Name: "nullable", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: false},
-				{Name: "case_sensitive", Type: flatbuf.TypeBool, GetTypeTable: boolTypeTable(), Nullable: false},
-				{Name: "searchable", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: false},
-				{Name: "unsigned_attribute", Type: flatbuf.TypeBool, GetTypeTable: boolTypeTable(), Nullable: true},
-				{Name: "fixed_prec_scale", Type: flatbuf.TypeBool, GetTypeTable: boolTypeTable(), Nullable: false},
-				{Name: "auto_increment", Type: flatbuf.TypeBool, GetTypeTable: boolTypeTable(), Nullable: true},
-				{Name: "local_type_name", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true},
-				{Name: "minimum_scale", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: true},
-				{Name: "maximum_scale", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: true},
-				{Name: "sql_data_type", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: false},
-				{Name: "datetime_subcode", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: true},
-				{Name: "num_prec_radix", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: true},
-				{Name: "interval_precision", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: true},
-			},
+			Fields:  serialize.SchemaRef.XdbcTypeInfo,
 		},
 		{
 			Command: &flight.CommandGetSqlInfo{Info: []uint32{0, 3}},
-			Fields: []field{
-				{Name: "info_name", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, false), Nullable: false},
-				{
-					Name: "value",
-					Type: flatbuf.TypeUnion,
-					GetTypeTable: unionTypeTable(
-						flatbuf.UnionModeDense,
-						[]field{
-							{Name: "string_value", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable()},
-							{Name: "bool_value", Type: flatbuf.TypeBool, GetTypeTable: boolTypeTable()},
-							{Name: "bigint_value", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(64, true)},
-							{Name: "int32_bitmask", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true)},
-							{
-								Name:         "string_list",
-								Type:         flatbuf.TypeList,
-								GetTypeTable: listTypeTable(field{Name: "item", Type: flatbuf.TypeUtf8, GetTypeTable: utf8TypeTable(), Nullable: true})},
-							{
-								Name: "int32_to_int32_list_map",
-								Type: flatbuf.TypeMap,
-								GetTypeTable: mapTypeTable(
-									field{Name: "key", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true)},
-									field{
-										Name:         "val",
-										Type:         flatbuf.TypeList,
-										GetTypeTable: listTypeTable(field{Name: "item", Type: flatbuf.TypeInt, GetTypeTable: intTypeTable(32, true), Nullable: true}),
-										Nullable:     true, // TODO: nullable?
-									},
-								),
-							},
-						},
-					), Nullable: false},
-			},
+			Fields:  serialize.SchemaRef.SqlInfo,
 		},
 	}
 
@@ -294,7 +154,7 @@ func init() {
 			Name: "GetFlightInfo/CommandStatementQuery",
 			ServerHandler: scenario.Handler{GetFlightInfo: func(ctx context.Context, fd *flight.FlightDescriptor) (*flight.FlightInfo, error) {
 				var cmd flight.CommandStatementQuery
-				if err := deserializeProtobufWrappedInAny(fd.Cmd, &cmd); err != nil {
+				if err := serialize.DeserializeProtobufWrappedInAny(fd.Cmd, &cmd); err != nil {
 					return nil, status.Errorf(codes.InvalidArgument, "failed to deserialize FlightDescriptor.Cmd: %s", err)
 				}
 
@@ -306,7 +166,7 @@ func init() {
 					return nil, status.Errorf(codes.InvalidArgument, "expected no TransactionID")
 				}
 
-				handle, err := createStatementQueryTicket([]byte(stmtQueryHandle))
+				handle, err := serialize.SerializeProtobufWrappedInAny(&flight.TicketStatementQuery{StatementHandle: []byte(stmtQueryHandle)})
 				if err != nil {
 					return nil, status.Errorf(codes.Internal, "failed to create Ticket: %s", err)
 				}
@@ -319,7 +179,7 @@ func init() {
 			}}},
 		scenario.ScenarioStep{Name: "DoGet/TicketStatementQuery", ServerHandler: scenario.Handler{DoGet: func(t *flight.Ticket, fs flight.FlightService_DoGetServer) error {
 			var cmd flight.TicketStatementQuery
-			if err := deserializeProtobufWrappedInAny(t.Ticket, &cmd); err != nil {
+			if err := serialize.DeserializeProtobufWrappedInAny(t.Ticket, &cmd); err != nil {
 				return status.Errorf(codes.InvalidArgument, "failed to deserialize Ticket.Ticket: %s", err)
 			}
 
@@ -327,9 +187,9 @@ func init() {
 				return status.Errorf(codes.InvalidArgument, "expected handle: %s, found: %s", stmtQueryHandle, cmd.GetStatementHandle())
 			}
 
-			return fs.Send(&flight.FlightData{DataHeader: buildFlatbufferSchema(queryFields)})
+			return fs.Send(&flight.FlightData{DataHeader: serialize.BuildFlatbufferSchema(serialize.SchemaExample.Query)})
 		}}},
-		scenario.ScenarioStep{Name: "GetSchema/CommandStatementQuery", ServerHandler: scenario.Handler{GetSchema: getSchemaFieldsForCommandFn(&flight.CommandStatementQuery{}, queryFields)}},
+		scenario.ScenarioStep{Name: "GetSchema/CommandStatementQuery", ServerHandler: scenario.Handler{GetSchema: getSchemaFieldsForCommandFn(&flight.CommandStatementQuery{}, serialize.SchemaExample.Query)}},
 
 		scenario.ScenarioStep{
 			Name: "DoPut/CommandStatementUpdate",
@@ -341,7 +201,7 @@ func init() {
 
 				desc := data.FlightDescriptor
 				var cmd flight.CommandStatementUpdate
-				if err := deserializeProtobufWrappedInAny(desc.Cmd, &cmd); err != nil {
+				if err := serialize.DeserializeProtobufWrappedInAny(desc.Cmd, &cmd); err != nil {
 					return status.Errorf(codes.InvalidArgument, "failed to deserialize FlightDescriptor.Cmd: %s", err)
 				}
 
@@ -369,7 +229,7 @@ func init() {
 			Name: "DoAction/ActionCreatePreparedStatementRequest",
 			ServerHandler: scenario.Handler{DoAction: func(a *flight.Action, fs flight.FlightService_DoActionServer) error {
 				var req flight.ActionCreatePreparedStatementRequest
-				if err := deserializeProtobufWrappedInAny(a.Body, &req); err != nil {
+				if err := serialize.DeserializeProtobufWrappedInAny(a.Body, &req); err != nil {
 					return status.Errorf(codes.InvalidArgument, "failed to deserialize Action.Body: %s", err)
 				}
 
@@ -381,7 +241,7 @@ func init() {
 					return status.Errorf(codes.InvalidArgument, "expected no TransactionID")
 				}
 
-				body, err := serializeProtobufWrappedInAny(&flight.ActionCreatePreparedStatementResult{PreparedStatementHandle: []byte(stmtPreparedQueryHandle)})
+				body, err := serialize.SerializeProtobufWrappedInAny(&flight.ActionCreatePreparedStatementResult{PreparedStatementHandle: []byte(stmtPreparedQueryHandle)})
 				if err != nil {
 					return status.Errorf(codes.Internal, "failed to ActionCreatePreparedStatementResult: %s", err)
 				}
@@ -401,21 +261,21 @@ func init() {
 					return status.Errorf(codes.Internal, "invalid stream, expected first message to be Schema: %s", err)
 				}
 
-				fields, ok := parseFlatbufferSchemaFields(msg)
+				fields, ok := serialize.ParseFlatbufferSchemaFields(msg)
 				if !ok {
 					return status.Errorf(codes.Internal, "failed to parse flatbuffer schema")
 				}
 
 				// TODO: maybe don't use tester here
 				t := tester.NewTester()
-				assertSchemaMatchesFields(t, fields, queryFields)
+				tester.AssertSchemaMatchesFields(t, fields, serialize.SchemaExample.Query)
 				if len(t.Errors()) > 0 {
 					return status.Errorf(codes.Internal, "flatbuffer schema mismatch: %s", errors.Join(t.Errors()...))
 				}
 
 				desc := data.FlightDescriptor
 				var cmd flight.CommandPreparedStatementQuery
-				if err := deserializeProtobufWrappedInAny(desc.Cmd, &cmd); err != nil {
+				if err := serialize.DeserializeProtobufWrappedInAny(desc.Cmd, &cmd); err != nil {
 					return status.Errorf(codes.InvalidArgument, "failed to deserialize FlightDescriptor.Cmd: %s", err)
 				}
 
@@ -434,7 +294,7 @@ func init() {
 			Name: "GetFlightInfo/CommandPreparedStatementQuery",
 			ServerHandler: scenario.Handler{GetFlightInfo: func(ctx context.Context, fd *flight.FlightDescriptor) (*flight.FlightInfo, error) {
 				var cmd flight.CommandPreparedStatementQuery
-				if err := deserializeProtobufWrappedInAny(fd.Cmd, &cmd); err != nil {
+				if err := serialize.DeserializeProtobufWrappedInAny(fd.Cmd, &cmd); err != nil {
 					return nil, status.Errorf(codes.InvalidArgument, "failed to deserialize FlightDescriptor.Cmd: %s", err)
 				}
 
@@ -452,7 +312,7 @@ func init() {
 			Name: "DoGet/CommandPreparedStatementQuery",
 			ServerHandler: scenario.Handler{DoGet: func(t *flight.Ticket, fs flight.FlightService_DoGetServer) error {
 				var cmd flight.CommandPreparedStatementQuery
-				if err := deserializeProtobufWrappedInAny(t.Ticket, &cmd); err != nil {
+				if err := serialize.DeserializeProtobufWrappedInAny(t.Ticket, &cmd); err != nil {
 					return status.Errorf(codes.InvalidArgument, "failed to deserialize Ticket.Ticket: %s", err)
 				}
 
@@ -460,13 +320,13 @@ func init() {
 					return status.Errorf(codes.InvalidArgument, "expected handle: %s, found: %s", stmtPreparedQueryHandle, cmd.GetPreparedStatementHandle())
 				}
 
-				return fs.Send(&flight.FlightData{DataHeader: buildFlatbufferSchema(queryFields)})
+				return fs.Send(&flight.FlightData{DataHeader: serialize.BuildFlatbufferSchema(serialize.SchemaExample.Query)})
 			}}},
 		scenario.ScenarioStep{
 			Name: "GetSchema/CommandPreparedStatementQuery",
 			ServerHandler: scenario.Handler{GetSchema: func(ctx context.Context, fd *flight.FlightDescriptor) (*flight.SchemaResult, error) {
 				var cmd flight.CommandPreparedStatementQuery
-				if err := deserializeProtobufWrappedInAny(fd.Cmd, &cmd); err != nil {
+				if err := serialize.DeserializeProtobufWrappedInAny(fd.Cmd, &cmd); err != nil {
 					return nil, status.Errorf(codes.InvalidArgument, "failed to deserialize FlightDescriptor.Cmd: %s", err)
 				}
 
@@ -474,13 +334,13 @@ func init() {
 					return nil, status.Errorf(codes.InvalidArgument, "expected handle: %s, found: %s", stmtPreparedQueryHandle, cmd.GetPreparedStatementHandle())
 				}
 
-				return &flight.SchemaResult{Schema: writeFlatbufferPayload(queryFields)}, nil
+				return &flight.SchemaResult{Schema: serialize.WriteFlatbufferPayload(serialize.SchemaExample.Query)}, nil
 			}}},
 		scenario.ScenarioStep{
 			Name: "DoAction/ActionClosePreparedStatementRequest",
 			ServerHandler: scenario.Handler{DoAction: func(a *flight.Action, fs flight.FlightService_DoActionServer) error {
 				var req flight.ActionClosePreparedStatementRequest
-				if err := deserializeProtobufWrappedInAny(a.Body, &req); err != nil {
+				if err := serialize.DeserializeProtobufWrappedInAny(a.Body, &req); err != nil {
 					return status.Errorf(codes.InvalidArgument, "failed to deserialize Action.Body: %s", err)
 				}
 
@@ -495,7 +355,7 @@ func init() {
 			Name: "DoAction/ActionCreatePreparedStatementRequest",
 			ServerHandler: scenario.Handler{DoAction: func(a *flight.Action, fs flight.FlightService_DoActionServer) error {
 				var req flight.ActionCreatePreparedStatementRequest
-				if err := deserializeProtobufWrappedInAny(a.Body, &req); err != nil {
+				if err := serialize.DeserializeProtobufWrappedInAny(a.Body, &req); err != nil {
 					return status.Errorf(codes.InvalidArgument, "failed to deserialize Action.Body: %s", err)
 				}
 
@@ -507,7 +367,7 @@ func init() {
 					return status.Errorf(codes.InvalidArgument, "expected no TransactionID")
 				}
 
-				body, err := serializeProtobufWrappedInAny(&flight.ActionCreatePreparedStatementResult{PreparedStatementHandle: []byte(stmtPreparedUpdateHandle)})
+				body, err := serialize.SerializeProtobufWrappedInAny(&flight.ActionCreatePreparedStatementResult{PreparedStatementHandle: []byte(stmtPreparedUpdateHandle)})
 				if err != nil {
 					return status.Errorf(codes.Internal, "failed to ActionCreatePreparedStatementResult: %s", err)
 				}
@@ -527,7 +387,7 @@ func init() {
 					return status.Errorf(codes.Internal, "invalid stream, expected first message to be Schema: %s", err)
 				}
 
-				fields, ok := parseFlatbufferSchemaFields(msg)
+				fields, ok := serialize.ParseFlatbufferSchemaFields(msg)
 				if !ok {
 					return status.Errorf(codes.Internal, "failed to parse flatbuffer schema")
 				}
@@ -538,7 +398,7 @@ func init() {
 
 				desc := data.FlightDescriptor
 				var cmd flight.CommandPreparedStatementUpdate
-				if err := deserializeProtobufWrappedInAny(desc.Cmd, &cmd); err != nil {
+				if err := serialize.DeserializeProtobufWrappedInAny(desc.Cmd, &cmd); err != nil {
 					return status.Errorf(codes.InvalidArgument, "failed to deserialize FlightDescriptor.Cmd: %s", err)
 				}
 
@@ -557,7 +417,7 @@ func init() {
 			Name: "DoAction/ActionClosePreparedStatementRequest",
 			ServerHandler: scenario.Handler{DoAction: func(a *flight.Action, fs flight.FlightService_DoActionServer) error {
 				var req flight.ActionClosePreparedStatementRequest
-				if err := deserializeProtobufWrappedInAny(a.Body, &req); err != nil {
+				if err := serialize.DeserializeProtobufWrappedInAny(a.Body, &req); err != nil {
 					return status.Errorf(codes.InvalidArgument, "failed to deserialize Action.Body: %s", err)
 				}
 
@@ -579,7 +439,7 @@ func init() {
 				////////////////////////////
 				for _, tc := range testcases {
 					// pack the command
-					desc, err := descForCommand(tc.Command)
+					desc, err := serialize.DescForCommand(tc.Command)
 					t.Require().NoError(err)
 
 					// submit query
@@ -594,10 +454,10 @@ func init() {
 					t.Require().NoError(err)
 
 					// validate first message is properly formatted schema message
-					requireStreamHeaderMatchesFields(t, stream, tc.Fields)
+					tester.RequireStreamHeaderMatchesFields(t, stream, tc.Fields)
 
 					// drain rest of stream
-					requireDrainStream(t, stream, func(t *tester.Tester, data *flight.FlightData) {
+					tester.RequireDrainStream(t, stream, func(t *tester.Tester, data *flight.FlightData) {
 						// no more schema messages
 						t.Assert().Contains(
 							[]flatbuf.MessageHeader{
@@ -613,14 +473,14 @@ func init() {
 					t.Require().NoError(err)
 
 					// expect schema to be serialized as full IPC stream Schema message
-					requireSchemaResultMatchesFields(t, res, tc.Fields)
+					tester.RequireSchemaResultMatchesFields(t, res, tc.Fields)
 				}
 
 				// ValidateStatementExecution
 				/////////////////////////////
 				{
 					{
-						desc, err := descForCommand(&flight.CommandStatementQuery{Query: stmtQuery})
+						desc, err := serialize.DescForCommand(&flight.CommandStatementQuery{Query: stmtQuery})
 						t.Require().NoError(err)
 
 						info, err := client.GetFlightInfo(ctx, desc)
@@ -633,8 +493,8 @@ func init() {
 						t.Require().NoError(err)
 
 						// validate result stream
-						requireStreamHeaderMatchesFields(t, stream, queryFields)
-						requireDrainStream(t, stream, func(t *tester.Tester, data *flight.FlightData) {
+						tester.RequireStreamHeaderMatchesFields(t, stream, serialize.SchemaExample.Query)
+						tester.RequireDrainStream(t, stream, func(t *tester.Tester, data *flight.FlightData) {
 							// no more schema messages
 							t.Assert().Contains(
 								[]flatbuf.MessageHeader{
@@ -649,11 +509,11 @@ func init() {
 						t.Require().NoError(err)
 
 						// expect schema to be serialized as full IPC stream Schema message
-						requireSchemaResultMatchesFields(t, res, queryFields)
+						tester.RequireSchemaResultMatchesFields(t, res, serialize.SchemaExample.Query)
 					}
 
 					{
-						desc, err := descForCommand(&flight.CommandStatementUpdate{Query: stmtUpdate})
+						desc, err := serialize.DescForCommand(&flight.CommandStatementUpdate{Query: stmtUpdate})
 						t.Require().NoError(err)
 
 						stream, err := client.DoPut(ctx)
@@ -677,10 +537,10 @@ func init() {
 				{
 					var prepareResult flight.ActionCreatePreparedStatementResult
 					{
-						prepareAction, err := packAction(
+						prepareAction, err := serialize.PackAction(
 							createPreparedStatementActionType,
 							&flight.ActionCreatePreparedStatementRequest{Query: stmtPreparedQuery},
-							serializeProtobufWrappedInAny,
+							serialize.SerializeProtobufWrappedInAny,
 						)
 						t.Require().NoError(err)
 
@@ -692,10 +552,10 @@ func init() {
 						result, err := stream.Recv()
 						t.Require().NoError(err)
 
-						t.Require().NoError(deserializeProtobufWrappedInAny(result.Body, &prepareResult))
+						t.Require().NoError(serialize.DeserializeProtobufWrappedInAny(result.Body, &prepareResult))
 
 						t.Require().Equal(stmtPreparedQueryHandle, string(prepareResult.GetPreparedStatementHandle()))
-						requireDrainStream(t, stream, nil)
+						tester.RequireDrainStream(t, stream, nil)
 					}
 
 					var doPutPreparedResult flight.DoPutPreparedStatementResult
@@ -703,10 +563,10 @@ func init() {
 						stream, err := client.DoPut(ctx)
 						t.Require().NoError(err)
 
-						desc, err := descForCommand(&flight.CommandPreparedStatementQuery{PreparedStatementHandle: prepareResult.GetPreparedStatementHandle()})
+						desc, err := serialize.DescForCommand(&flight.CommandPreparedStatementQuery{PreparedStatementHandle: prepareResult.GetPreparedStatementHandle()})
 						t.Require().NoError(err)
 
-						t.Require().NoError(stream.Send(&flight.FlightData{FlightDescriptor: desc, DataHeader: buildFlatbufferSchema(queryFields)}))
+						t.Require().NoError(stream.Send(&flight.FlightData{FlightDescriptor: desc, DataHeader: serialize.BuildFlatbufferSchema(serialize.SchemaExample.Query)}))
 						t.Require().NoError(stream.CloseSend())
 
 						putResult, err := stream.Recv()
@@ -723,7 +583,7 @@ func init() {
 						ticket          *flight.Ticket
 					)
 					{
-						desc, err := descForCommand(&flight.CommandPreparedStatementQuery{PreparedStatementHandle: doPutPreparedResult.GetPreparedStatementHandle()})
+						desc, err := serialize.DescForCommand(&flight.CommandPreparedStatementQuery{PreparedStatementHandle: doPutPreparedResult.GetPreparedStatementHandle()})
 						t.Require().NoError(err)
 
 						info, err := client.GetFlightInfo(ctx, desc)
@@ -740,8 +600,8 @@ func init() {
 						t.Require().NoError(err)
 
 						// validate result stream
-						requireStreamHeaderMatchesFields(t, stream, queryFields)
-						requireDrainStream(t, stream, func(t *tester.Tester, data *flight.FlightData) {
+						tester.RequireStreamHeaderMatchesFields(t, stream, serialize.SchemaExample.Query)
+						tester.RequireDrainStream(t, stream, func(t *tester.Tester, data *flight.FlightData) {
 							// no more schema messages
 							t.Assert().Contains(
 								[]flatbuf.MessageHeader{
@@ -757,14 +617,14 @@ func init() {
 						schema, err := client.GetSchema(ctx, descPutPrepared)
 						t.Require().NoError(err)
 
-						requireSchemaResultMatchesFields(t, schema, queryFields)
+						tester.RequireSchemaResultMatchesFields(t, schema, serialize.SchemaExample.Query)
 					}
 
 					{
-						action, err := packAction(
+						action, err := serialize.PackAction(
 							closePreparedStatementActionType,
 							&flight.ActionClosePreparedStatementRequest{PreparedStatementHandle: []byte(stmtPreparedQueryHandle)},
-							serializeProtobufWrappedInAny,
+							serialize.SerializeProtobufWrappedInAny,
 						)
 						t.Require().NoError(err)
 
@@ -772,14 +632,14 @@ func init() {
 						t.Require().NoError(err)
 
 						t.Require().NoError(stream.CloseSend())
-						requireDrainStream(t, stream, nil)
+						tester.RequireDrainStream(t, stream, nil)
 					}
 
 					{
-						action, err := packAction(
+						action, err := serialize.PackAction(
 							createPreparedStatementActionType,
 							&flight.ActionCreatePreparedStatementRequest{Query: stmtPreparedUpdate},
-							serializeProtobufWrappedInAny,
+							serialize.SerializeProtobufWrappedInAny,
 						)
 						t.Require().NoError(err)
 
@@ -792,20 +652,20 @@ func init() {
 						t.Require().NoError(err)
 
 						var actionResult flight.ActionCreatePreparedStatementResult
-						t.Require().NoError(deserializeProtobufWrappedInAny(result.Body, &actionResult))
+						t.Require().NoError(serialize.DeserializeProtobufWrappedInAny(result.Body, &actionResult))
 
 						t.Require().Equal(stmtPreparedUpdateHandle, string(actionResult.GetPreparedStatementHandle()))
-						requireDrainStream(t, stream, nil)
+						tester.RequireDrainStream(t, stream, nil)
 					}
 
 					{
 						stream, err := client.DoPut(ctx)
 						t.Require().NoError(err)
 
-						desc, err := descForCommand(&flight.CommandPreparedStatementUpdate{PreparedStatementHandle: []byte(stmtPreparedUpdateHandle)})
+						desc, err := serialize.DescForCommand(&flight.CommandPreparedStatementUpdate{PreparedStatementHandle: []byte(stmtPreparedUpdateHandle)})
 						t.Require().NoError(err)
 
-						t.Require().NoError(stream.Send(&flight.FlightData{FlightDescriptor: desc, DataHeader: buildFlatbufferSchema(nil)}))
+						t.Require().NoError(stream.Send(&flight.FlightData{FlightDescriptor: desc, DataHeader: serialize.BuildFlatbufferSchema(nil)}))
 						t.Require().NoError(stream.CloseSend())
 
 						putResult, err := stream.Recv()
@@ -818,10 +678,10 @@ func init() {
 					}
 
 					{
-						action, err := packAction(
+						action, err := serialize.PackAction(
 							closePreparedStatementActionType,
 							&flight.ActionClosePreparedStatementRequest{PreparedStatementHandle: []byte(stmtPreparedUpdateHandle)},
-							serializeProtobufWrappedInAny,
+							serialize.SerializeProtobufWrappedInAny,
 						)
 						t.Require().NoError(err)
 
@@ -829,288 +689,12 @@ func init() {
 						t.Require().NoError(err)
 
 						t.Require().NoError(stream.CloseSend())
-						requireDrainStream(t, stream, nil)
+						tester.RequireDrainStream(t, stream, nil)
 					}
 				}
 			},
 		},
 	)
-}
-
-type typeTableFn func(*flatbuffers.Builder) (flatbuffers.UOffsetT, flatbuffers.UOffsetT)
-
-type field struct {
-	Name         string
-	Type         flatbuf.Type
-	Nullable     bool
-	Metadata     map[string]string
-	GetTypeTable typeTableFn
-}
-
-func utf8TypeTable() typeTableFn {
-	return func(b *flatbuffers.Builder) (flatbuffers.UOffsetT, flatbuffers.UOffsetT) {
-		flatbuf.Utf8Start(b)
-		return flatbuf.Utf8End(b), 0
-	}
-}
-
-func binaryTypeTable() typeTableFn {
-	return func(b *flatbuffers.Builder) (flatbuffers.UOffsetT, flatbuffers.UOffsetT) {
-		flatbuf.BinaryStart(b)
-		return flatbuf.BinaryEnd(b), 0
-	}
-}
-
-func intTypeTable(bitWidth int32, isSigned bool) typeTableFn {
-	return func(b *flatbuffers.Builder) (flatbuffers.UOffsetT, flatbuffers.UOffsetT) {
-		flatbuf.IntStart(b)
-		flatbuf.IntAddBitWidth(b, bitWidth)
-		flatbuf.IntAddIsSigned(b, isSigned)
-		return flatbuf.IntEnd(b), 0
-	}
-}
-
-func boolTypeTable() typeTableFn {
-	return func(b *flatbuffers.Builder) (flatbuffers.UOffsetT, flatbuffers.UOffsetT) {
-		flatbuf.BoolStart(b)
-		return flatbuf.BoolEnd(b), 0
-	}
-}
-
-func listTypeTable(child field) typeTableFn {
-	return func(b *flatbuffers.Builder) (flatbuffers.UOffsetT, flatbuffers.UOffsetT) {
-		childOffset := buildFlatbufferField(b, child)
-
-		flatbuf.ListStart(b)
-		listOffset := flatbuf.ListEnd(b)
-
-		flatbuf.FieldStartChildrenVector(b, 1)
-		b.PrependUOffsetT(childOffset)
-		childVecOffset := b.EndVector(1)
-
-		return listOffset, childVecOffset
-	}
-}
-
-func structTypeTable(children []field) typeTableFn {
-	return func(b *flatbuffers.Builder) (flatbuffers.UOffsetT, flatbuffers.UOffsetT) {
-		nChildren := len(children)
-		childOffsets := make([]flatbuffers.UOffsetT, nChildren)
-		for i, child := range children {
-			childOffsets[i] = buildFlatbufferField(b, child)
-		}
-
-		flatbuf.Struct_Start(b)
-		for i := nChildren - 1; i >= 0; i-- {
-			b.PrependUOffsetT(childOffsets[i])
-		}
-		structOffset := flatbuf.Struct_End(b)
-
-		flatbuf.FieldStartChildrenVector(b, nChildren)
-		for i := nChildren - 1; i >= 0; i-- {
-			b.PrependUOffsetT(childOffsets[i])
-		}
-		childVecOffset := b.EndVector(nChildren)
-
-		return structOffset, childVecOffset
-	}
-}
-
-func mapTypeTable(key, val field) typeTableFn {
-	return func(b *flatbuffers.Builder) (flatbuffers.UOffsetT, flatbuffers.UOffsetT) {
-		childOffset := buildFlatbufferField(
-			b,
-			field{
-				Name:         "entries",
-				Type:         flatbuf.TypeStruct_,
-				GetTypeTable: structTypeTable([]field{key, val}),
-			},
-		)
-
-		flatbuf.MapStart(b)
-		mapOffset := flatbuf.MapEnd(b)
-
-		flatbuf.FieldStartChildrenVector(b, 1)
-		b.PrependUOffsetT(childOffset)
-		childVecOffset := b.EndVector(1)
-
-		return mapOffset, childVecOffset
-	}
-}
-
-func unionTypeTable(mode flatbuf.UnionMode, children []field) typeTableFn {
-	return func(b *flatbuffers.Builder) (flatbuffers.UOffsetT, flatbuffers.UOffsetT) {
-		childOffsets := make([]flatbuffers.UOffsetT, len(children))
-		for i, child := range children {
-			childOffsets[i] = buildFlatbufferField(b, child)
-		}
-
-		flatbuf.UnionStartTypeIdsVector(b, len(children))
-		for i := len(children) - 1; i >= 0; i-- {
-			b.PlaceInt32(int32(i))
-		}
-		typeIDVecOffset := b.EndVector(len(children))
-
-		flatbuf.UnionStart(b)
-		flatbuf.UnionAddMode(b, mode)
-		flatbuf.UnionAddTypeIds(b, typeIDVecOffset)
-		unionOffset := flatbuf.UnionEnd(b)
-
-		flatbuf.FieldStartChildrenVector(b, len(children))
-		for i := len(children) - 1; i >= 0; i-- {
-			b.PrependUOffsetT(childOffsets[i])
-		}
-		childVecOffset := b.EndVector(len(children))
-
-		return unionOffset, childVecOffset
-	}
-}
-
-func matchFieldByName(fields []flatbuf.Field, name string) (flatbuf.Field, bool) {
-	for _, f := range fields {
-		fieldName := string(f.Name())
-		if fieldName == name {
-			return f, true
-		}
-	}
-	return flatbuf.Field{}, false
-}
-
-func writeFlatbufferPayload(fields []field) []byte {
-	schema := buildFlatbufferSchema(fields)
-	size := uint32(len(schema))
-
-	res := make([]byte, 8+size)
-	res[0] = 255
-	res[1] = 255
-	res[2] = 255
-	res[3] = 255
-	binary.LittleEndian.PutUint32(res[4:], size)
-	copy(res[8:], schema)
-
-	return res
-}
-
-func buildFlatbufferSchema(fields []field) []byte {
-	b := flatbuffers.NewBuilder(1024)
-
-	fieldOffsets := make([]flatbuffers.UOffsetT, len(fields))
-	for i, f := range fields {
-		fieldOffsets[len(fields)-i-1] = buildFlatbufferField(b, f)
-	}
-
-	flatbuf.SchemaStartFieldsVector(b, len(fields))
-
-	for _, f := range fieldOffsets {
-		b.PrependUOffsetT(f)
-	}
-
-	fieldsFB := b.EndVector(len(fields))
-
-	flatbuf.SchemaStart(b)
-	flatbuf.SchemaAddFields(b, fieldsFB)
-	headerOffset := flatbuf.SchemaEnd(b)
-
-	flatbuf.MessageStart(b)
-	flatbuf.MessageAddVersion(b, flatbuf.MetadataVersionV5)
-	flatbuf.MessageAddHeaderType(b, flatbuf.MessageHeaderSchema)
-	flatbuf.MessageAddHeader(b, headerOffset)
-	msg := flatbuf.MessageEnd(b)
-
-	b.Finish(msg)
-
-	return b.FinishedBytes()
-}
-
-func buildFlatbufferField(b *flatbuffers.Builder, f field) flatbuffers.UOffsetT {
-	nameOffset := b.CreateString(f.Name)
-	typOffset, childrenOffset := f.GetTypeTable(b)
-
-	var kvOffsets []flatbuffers.UOffsetT
-	for k, v := range f.Metadata {
-		kk := b.CreateString(k)
-		vv := b.CreateString(v)
-		flatbuf.KeyValueStart(b)
-		flatbuf.KeyValueAddKey(b, kk)
-		flatbuf.KeyValueAddValue(b, vv)
-		kvOffsets = append(kvOffsets, flatbuf.KeyValueEnd(b))
-	}
-
-	var metadataOffset flatbuffers.UOffsetT
-	if len(kvOffsets) > 0 {
-		flatbuf.FieldStartCustomMetadataVector(b, len(kvOffsets))
-		for i := len(kvOffsets) - 1; i >= 0; i-- {
-			b.PrependUOffsetT(kvOffsets[i])
-		}
-		metadataOffset = b.EndVector(len(kvOffsets))
-	}
-
-	flatbuf.FieldStart(b)
-	flatbuf.FieldAddName(b, nameOffset)
-	flatbuf.FieldAddTypeType(b, f.Type)
-	flatbuf.FieldAddType(b, typOffset)
-	flatbuf.FieldAddChildren(b, childrenOffset)
-	flatbuf.FieldAddCustomMetadata(b, metadataOffset)
-	flatbuf.FieldAddNullable(b, f.Nullable)
-	return flatbuf.FieldEnd(b)
-}
-
-func parseFlatbufferSchemaFields(msg *flatbuf.Message) ([]flatbuf.Field, bool) {
-	var schema flatbuf.Schema
-	table := schema.Table()
-	if ok := msg.Header(&table); !ok {
-		return nil, false
-	}
-	schema.Init(table.Bytes, table.Pos)
-
-	fields := make([]flatbuf.Field, schema.FieldsLength())
-	for i := range fields {
-		var field flatbuf.Field
-		if ok := schema.Fields(&field, i); !ok {
-			return nil, false
-		}
-		fields[i] = field
-	}
-	return fields, true
-}
-
-func extractFlatbufferPayload(b []byte) []byte {
-	b = consumeContinuationIndicator(b)
-	b, size := consumeMetadataSize(b)
-	return b[:size]
-}
-
-func consumeMetadataSize(b []byte) ([]byte, int32) {
-	size := int32(binary.LittleEndian.Uint32(b[:4]))
-	return b[4:], size
-}
-
-func consumeContinuationIndicator(b []byte) []byte {
-	indicator := []byte{255, 255, 255, 255}
-	for i, v := range indicator {
-		if b[i] != v {
-			// indicator not found
-			return b
-		}
-	}
-	// indicator found, truncate leading bytes
-	return b[4:]
-}
-
-func descForCommand(cmd proto.Message) (*flight.FlightDescriptor, error) {
-	var any anypb.Any
-	if err := any.MarshalFrom(cmd); err != nil {
-		return nil, err
-	}
-
-	data, err := proto.Marshal(&any)
-	if err != nil {
-		return nil, err
-	}
-	return &flight.FlightDescriptor{
-		Type: flight.FlightDescriptor_CMD,
-		Cmd:  data,
-	}, nil
 }
 
 func echoFlightInfo(ctx context.Context, fd *flight.FlightDescriptor) (*flight.FlightInfo, error) {
@@ -1121,164 +705,30 @@ func echoFlightInfo(ctx context.Context, fd *flight.FlightDescriptor) (*flight.F
 	}, nil
 }
 
-func doGetFieldsForCommandFn(cmd proto.Message, fields []field) func(t *flight.Ticket, fs flight.FlightService_DoGetServer) error {
+func doGetFieldsForCommandFn(cmd proto.Message, fields []serialize.Field) func(t *flight.Ticket, fs flight.FlightService_DoGetServer) error {
 	return func(t *flight.Ticket, fs flight.FlightService_DoGetServer) error {
 		cmd := proto.Clone(cmd)
 		proto.Reset(cmd)
 
-		if err := deserializeProtobufWrappedInAny(t.Ticket, cmd); err != nil {
+		if err := serialize.DeserializeProtobufWrappedInAny(t.Ticket, cmd); err != nil {
 			return status.Errorf(codes.InvalidArgument, "failed to deserialize Ticket.Ticket: %s", err)
 		}
 
-		return fs.Send(&flight.FlightData{DataHeader: buildFlatbufferSchema(fields)})
+		return fs.Send(&flight.FlightData{DataHeader: serialize.BuildFlatbufferSchema(fields)})
 	}
 }
 
-func getSchemaFieldsForCommandFn(cmd proto.Message, fields []field) func(ctx context.Context, fd *flight.FlightDescriptor) (*flight.SchemaResult, error) {
+func getSchemaFieldsForCommandFn(cmd proto.Message, fields []serialize.Field) func(ctx context.Context, fd *flight.FlightDescriptor) (*flight.SchemaResult, error) {
 	return func(ctx context.Context, fd *flight.FlightDescriptor) (*flight.SchemaResult, error) {
 		cmd := proto.Clone(cmd)
 		proto.Reset(cmd)
 
-		if err := deserializeProtobufWrappedInAny(fd.Cmd, cmd); err != nil {
+		if err := serialize.DeserializeProtobufWrappedInAny(fd.Cmd, cmd); err != nil {
 			return nil, status.Errorf(codes.InvalidArgument, "failed to deserialize FlightDescriptor.Cmd: %s", err)
 		}
 
-		schema := writeFlatbufferPayload(fields)
+		schema := serialize.WriteFlatbufferPayload(fields)
 
 		return &flight.SchemaResult{Schema: schema}, nil
 	}
-}
-
-func deserializeProtobufWrappedInAny(b []byte, dst proto.Message) error {
-	var anycmd anypb.Any
-	if err := proto.Unmarshal(b, &anycmd); err != nil {
-		return fmt.Errorf("unable to unmarshal payload to proto.Any: %s", err)
-	}
-
-	if err := anycmd.UnmarshalTo(dst); err != nil {
-		return fmt.Errorf("unable to unmarshal proto.Any: %s", err)
-	}
-
-	return nil
-}
-
-func deserializeProtobuf(b []byte, dst proto.Message) error {
-	if err := proto.Unmarshal(b, dst); err != nil {
-		return fmt.Errorf("unable to unmarshal protobuf payload: %s", err)
-	}
-
-	return nil
-}
-
-func createStatementQueryTicket(handle []byte) ([]byte, error) {
-	query := &flight.TicketStatementQuery{StatementHandle: handle}
-
-	var ticket anypb.Any
-	if err := ticket.MarshalFrom(query); err != nil {
-		return nil, fmt.Errorf("unable to marshal ticket proto to proto.Any: %s", err)
-	}
-
-	b, err := proto.Marshal(&ticket)
-	if err != nil {
-		return nil, fmt.Errorf("unable to marshal proto.Any to bytes: %s", err)
-	}
-
-	return b, nil
-}
-
-func serializeProtobufWrappedInAny(msg proto.Message) ([]byte, error) {
-	var anycmd anypb.Any
-	if err := anycmd.MarshalFrom(msg); err != nil {
-		return nil, fmt.Errorf("unable to marshal proto message to proto.Any: %s", err)
-	}
-
-	b, err := proto.Marshal(&anycmd)
-	if err != nil {
-		return nil, fmt.Errorf("unable to marshal proto.Any to bytes: %s", err)
-	}
-
-	return b, nil
-}
-
-func serializeProtobuf(msg proto.Message) ([]byte, error) {
-	b, err := proto.Marshal(msg)
-	if err != nil {
-		return nil, fmt.Errorf("unable to marshal protobuf message to bytes: %s", err)
-	}
-
-	return b, nil
-}
-
-func packAction(actionType string, msg proto.Message, serialize func(proto.Message) ([]byte, error)) (*flight.Action, error) {
-	var action flight.Action
-	body, err := serialize(msg)
-	if err != nil {
-		return nil, fmt.Errorf("failed to serialize action body: %s", err)
-	}
-
-	action.Type = actionType
-	action.Body = body
-
-	return &action, nil
-}
-
-func requireStreamHeaderMatchesFields[T interface {
-	Recv() (*flight.FlightData, error)
-}](t *tester.Tester, stream T, expectedFields []field) {
-
-	data, err := stream.Recv()
-	t.Require().NoError(err)
-
-	msg := flatbuf.GetRootAsMessage(data.DataHeader, 0)
-	t.Require().Equal(flatbuf.MessageHeaderSchema, msg.HeaderType())
-
-	fields, ok := parseFlatbufferSchemaFields(msg)
-	t.Require().True(ok)
-	t.Require().Len(fields, len(expectedFields))
-
-	assertSchemaMatchesFields(t, fields, expectedFields)
-}
-
-func assertSchemaMatchesFields(t *tester.Tester, fields []flatbuf.Field, expectedFields []field) {
-	for _, expectedField := range expectedFields {
-		field, found := matchFieldByName(fields, expectedField.Name)
-		t.Assert().Truef(found, "no matching field with expected name \"%s\" found in flatbuffer schema", expectedField.Name)
-		if !found {
-			continue
-		}
-
-		t.Assert().Equal(expectedField.Name, string(field.Name()))
-		t.Assert().Equal(expectedField.Type, field.TypeType())
-		t.Assert().Equal(expectedField.Nullable, field.Nullable())
-		// TODO: metadata?
-	}
-}
-
-func requireDrainStream[E any, S interface {
-	Recv() (E, error)
-}](t *tester.Tester, stream S, eval func(*tester.Tester, E)) {
-	for {
-		data, err := stream.Recv()
-		if err == io.EOF {
-			break
-		}
-		t.Require().NoError(err)
-
-		if eval != nil {
-			eval(t, data)
-		}
-	}
-}
-
-func requireSchemaResultMatchesFields(t *tester.Tester, res *flight.SchemaResult, expectedFields []field) {
-	metadata := extractFlatbufferPayload(res.Schema)
-
-	msg := flatbuf.GetRootAsMessage(metadata, 0)
-	t.Require().Equal(flatbuf.MessageHeaderSchema, msg.HeaderType())
-
-	fields, ok := parseFlatbufferSchemaFields(msg)
-	t.Require().True(ok)
-	t.Require().Len(fields, len(expectedFields))
-
-	assertSchemaMatchesFields(t, fields, expectedFields)
 }
